Outer product work in progress

+++ hls_v2/mat_mult_sparse_outer.h
@@ -8,18 +8,18 @@
 #define PE 8
 
 void mat_mult(
-    const uint8_t*  A_row_ind, 
-    const uint8_t*  A_col_ptr, 
-    const uint32_t* A_val, 
-    const uint16_t  A_nnz, 
-    const uint8_t*  B_row_ptr, 
-    const uint8_t*  B_col_ind, 
-    const uint32_t* B_val, 
-    const uint16_t  B_nnz, 
-    uint8_t*        C_row_ind,
-    uint8_t*        C_col_ind,
-    uint32_t*       C_val,
-    uint16_t*       C_nnz
+    const uint32_t*  A_row_ind,
+    const uint32_t*  A_col_ptr,
+    const uint32_t*  A_val,
+    const uint32_t   A_nnz,
+    const uint32_t*  B_row_ptr,
+    const uint32_t*  B_col_ind,
+    const uint32_t*  B_val,
+    const uint32_t   B_nnz,
+    uint32_t*        C_row_ind,
+    uint32_t*        C_col_ind,
+    uint32_t*        C_val,
+    uint32_t*        C_nnz
 );
 
 #endif
